backend/services/mqtt_bridge/requirements.txt
fastapi
paho-mqtt
httpx
pydantic
